+++ Source/Core/Properties+Style.swift
@@ -8,7 +8,7 @@
 
 extension Properties.Style {
     
-    public static func style<T>(for type: T.Type) -> Property<(T) -> Void> {
-        return Property<(T) -> Void>(defaultValue: { _ in })
+    public static func style<T>(for type: T.Type, defaultValue: @escaping (T) -> Void = { _ in }) -> Property<(T) -> Void> {
+        return Property<(T) -> Void>(defaultValue: defaultValue)
     }
 }